@@ -1,36 +1,37 @@
 #include "forms/shared/validators/date_validator.hpp"
 
 #include <cctype>
+#include <string>
+#include <string_view>
 
 namespace forms::shared
 {
 
-  form::ValidationResult DateValidator::ValidateNumericString(const std::string &str, const std::string &field_name) const
+  form::ValidationResult DateValidator::ValidateNumericString(std::string_view str, std::string_view field_name) const
   {
     for (char c : str)
     {
       if (!std::isdigit(c))
       {
-        return form::ValidationResult::Invalid(field_name + " must be numeric");
+        return form::ValidationResult::Invalid(std::string(field_name) + " must be numeric");
       }
     }
     return form::ValidationResult::Valid();
   }
 
-  form::ValidationResult DateValidator::ValidateDay(const std::string &day_str) const
+  form::ValidationResult DateValidator::ValidateDay(std::string_view day_str) const
   {
     if (day_str.empty() || day_str.length() > 2)
     {
       return form::ValidationResult::Invalid("Day must be 1-2 digits");
     }
 
-    auto numeric_result = ValidateNumericString(day_str, "Day");
-    if (!numeric_result.is_valid)
+    if (auto numeric_result = ValidateNumericString(day_str, "Day"); !numeric_result.is_valid)
     {
       return numeric_result;
     }
 
-    int day = std::stoi(day_str);
+    int day = std::stoi(std::string(day_str));
     if (day < 1 || day > 31)
     {
       return form::ValidationResult::Invalid("Day must be between 1 and 31");
@@ -39,20 +40,19 @@ namespace forms::shared
     return form::ValidationResult::Valid();
   }
 
-  form::ValidationResult DateValidator::ValidateMonth(const std::string &month_str) const
+  form::ValidationResult DateValidator::ValidateMonth(std::string_view month_str) const
   {
     if (month_str.empty() || month_str.length() > 2)
     {
       return form::ValidationResult::Invalid("Month must be 1-2 digits");
     }
 
-    auto numeric_result = ValidateNumericString(month_str, "Month");
-    if (!numeric_result.is_valid)
+    if (auto numeric_result = ValidateNumericString(month_str, "Month"); !numeric_result.is_valid)
     {
       return numeric_result;
     }
 
-    int month = std::stoi(month_str);
+    int month = std::stoi(std::string(month_str));
     if (month < 1 || month > 12)
     {
       return form::ValidationResult::Invalid("Month must be between 1 and 12");
@@ -61,20 +61,19 @@ namespace forms::shared
     return form::ValidationResult::Valid();
   }
 
-  form::ValidationResult DateValidator::ValidateYear(const std::string &year_str) const
+  form::ValidationResult DateValidator::ValidateYear(std::string_view year_str) const
   {
     if (year_str.length() != 4)
     {
       return form::ValidationResult::Invalid("Year must be 4 digits");
     }
 
-    auto numeric_result = ValidateNumericString(year_str, "Year");
-    if (!numeric_result.is_valid)
+    if (auto numeric_result = ValidateNumericString(year_str, "Year"); !numeric_result.is_valid)
     {
       return numeric_result;
     }
 
-    int year = std::stoi(year_str);
+    int year = std::stoi(std::string(year_str));
     if (year < 1900 || year > 2100)
     {
       return form::ValidationResult::Invalid("Year must be between 1900 and 2100");
@@ -83,7 +82,7 @@ namespace forms::shared
     return form::ValidationResult::Valid();
   }
 
-  form::ValidationResult DateValidator::ValidateDMYFormat(const std::string &value) const
+  form::ValidationResult DateValidator::ValidateDMYFormat(std::string_view value) const
   {
     size_t first_slash = value.find('/');
     size_t second_slash = value.find('/', first_slash + 1);
@@ -93,24 +92,21 @@ namespace forms::shared
       return form::ValidationResult::Invalid("Invalid date format. Expected DD/MM/YYYY");
     }
 
-    std::string day_str = value.substr(0, first_slash);
-    std::string month_str = value.substr(first_slash + 1, second_slash - first_slash - 1);
-    std::string year_str = value.substr(second_slash + 1);
+    std::string_view day_str = value.substr(0, first_slash);
+    std::string_view month_str = value.substr(first_slash + 1, second_slash - first_slash - 1);
+    std::string_view year_str = value.substr(second_slash + 1);
 
-    auto day_result = ValidateDay(day_str);
-    if (!day_result.is_valid)
+    if (auto day_result = ValidateDay(day_str); !day_result.is_valid)
     {
       return day_result;
     }
 
-    auto month_result = ValidateMonth(month_str);
-    if (!month_result.is_valid)
+    if (auto month_result = ValidateMonth(month_str); !month_result.is_valid)
     {
       return month_result;
     }
 
-    auto year_result = ValidateYear(year_str);
-    if (!year_result.is_valid)
+    if (auto year_result = ValidateYear(year_str); !year_result.is_valid)
     {
       return year_result;
     }
@@ -118,7 +114,7 @@ namespace forms::shared
     return form::ValidationResult::Valid();
   }
 
-  form::ValidationResult DateValidator::ValidateYMDFormat(const std::string &value) const
+  form::ValidationResult DateValidator::ValidateYMDFormat(std::string_view value) const
   {
     size_t first_dash = value.find('-');
     size_t second_dash = value.find('-', first_dash + 1);
@@ -128,24 +124,21 @@ namespace forms::shared
       return form::ValidationResult::Invalid("Invalid date format. Expected YYYY-MM-DD");
     }
 
-    std::string year_str = value.substr(0, first_dash);
-    std::string month_str = value.substr(first_dash + 1, second_dash - first_dash - 1);
-    std::string day_str = value.substr(second_dash + 1);
+    std::string_view year_str = value.substr(0, first_dash);
+    std::string_view month_str = value.substr(first_dash + 1, second_dash - first_dash - 1);
+    std::string_view day_str = value.substr(second_dash + 1);
 
-    auto year_result = ValidateYear(year_str);
-    if (!year_result.is_valid)
+    if (auto year_result = ValidateYear(year_str); !year_result.is_valid)
     {
       return year_result;
     }
 
-    auto month_result = ValidateMonth(month_str);
-    if (!month_result.is_valid)
+    if (auto month_result = ValidateMonth(month_str); !month_result.is_valid)
     {
       return month_result;
     }
 
-    auto day_result = ValidateDay(day_str);
-    if (!day_result.is_valid)
+    if (auto day_result = ValidateDay(day_str); !day_result.is_valid)
     {
       return day_result;
     }
@@ -161,16 +154,14 @@ namespace forms::shared
       return form::ValidationResult::Invalid("Date cannot be empty");
     }
 
-    size_t first_slash = value.find('/');
-    if (first_slash != std::string::npos)
+    if (size_t first_slash = value.find('/'); first_slash != std::string::npos)
     {
-      return ValidateDMYFormat(value);
+      return ValidateDMYFormat(std::string_view(value));
     }
 
-    size_t first_dash = value.find('-');
-    if (first_dash != std::string::npos)
+    if (size_t first_dash = value.find('-'); first_dash != std::string::npos)
     {
-      return ValidateYMDFormat(value);
+      return ValidateYMDFormat(std::string_view(value));
     }
 
     return form::ValidationResult::Invalid("Invalid date format. Expected DD/MM/YYYY or YYYY-MM-DD");
